Androidでコンパイルが通らないところの修正。 Android版仮ヘッダファイル追加・修正。 DrawDeviceの一部に環境依存の型が使われていたので修正。

+++ base/SystemIntf.cpp
@@ -103,7 +103,7 @@ TJS_END_NATIVE_CONSTRUCTOR_DECL(/*TJS class name*/System)
 //----------------------------------------------------------------------
 TJS_BEGIN_NATIVE_METHOD_DECL(/*func. name*/terminate)
 {
-	int code = numparams > 0 ? *param[0] : 0;
+	int code = numparams > 0 ? static_cast<int>(*param[0]) : 0;
 	TVPTerminateAsync(code);
 
 	return TJS_S_OK;
@@ -114,7 +114,7 @@ TJS_BEGIN_NATIVE_METHOD_DECL(/*func. name*/exit)
 {
 	// this method does not return
 
-	int code = numparams > 0 ? *param[0] : 0;
+	int code = numparams > 0 ? static_cast<int>(*param[0]) : 0;
 	TVPTerminateSync(code);
 
 	return TJS_S_OK;
